feature: moves `jsonStringToMapStringInterface` and `mapStringInterfaceToJSONString` to helpers.
