+++ fusionauth/helpers.go
@@ -5,8 +5,10 @@ import (
 	"fmt"
 	"net/http"
 	"reflect"
+	"strings"
 
 	"github.com/FusionAuth/go-client/pkg/fusionauth"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -56,4 +58,56 @@ func injectSchemaChanges(schemaToEdit, schemaEdits *schema.Resource) *schema.Res
 	}
 
 	return schemaToEdit
+}
+
+// jsonStringToMapStringInterface takes in a json encoded string and transforms
+// the data to a map[string]interface{} to comply to the expected type for the
+// fusionauth client.
+func jsonStringToMapStringInterface(fieldName string, data *schema.ResourceData) (out map[string]interface{}, diags diag.Diagnostics) {
+	in := data.Get(fieldName).(string)
+	out = map[string]interface{}{}
+	if strings.TrimSpace(in) == "" {
+		return out, nil
+	}
+
+	if err := json.Unmarshal([]byte(in), &out); err != nil {
+		fmt.Printf("jsonStringToMapStringInterface: %s", err)
+		diags = append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  fmt.Sprintf("Unable to transform %s to expected type", fieldName),
+			Detail: fmt.Sprintf(
+				"Error unmarshalling %s from an expected JSON encoded string to map[string]interface{}.\n"+
+					"Please make sure you have wrapped your HCL with jsonencode."+
+					"For example, 'jsonencode({ hello = \"world\" })'.\n\n"+
+					"error: %s\n",
+				fieldName, err,
+			),
+		})
+	}
+
+	return
+}
+
+// mapStringInterfaceToJSONString transforms a map[string]interface{} to a JSON
+// string.
+func mapStringInterfaceToJSONString(fieldName string, in map[string]interface{}) (out string, diags diag.Diagnostics) {
+	if len(in) == 0 {
+		return "", nil
+	}
+
+	outBytes, err := json.MarshalIndent(in, "", "  ")
+	if err != nil {
+		diags = append(diags, diag.Diagnostic{
+			Severity: diag.Error,
+			Summary:  fmt.Sprintf("Unable to transform %s to expected type", fieldName),
+			Detail: fmt.Sprintf(
+				"Error marshalling %s from a map[string]interface{} to a JSON string.\n"+
+					"error: %s\n",
+				fieldName, err,
+			),
+		})
+		return
+	}
+
+	return string(outBytes), nil
 }
