+++ fusionauth/resource_fusionauth_user.go
@@ -2,7 +2,6 @@ package fusionauth
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"net/http"
 	"strings"
@@ -118,21 +117,21 @@ func deleteUser(_ context.Context, data *schema.ResourceData, i interface{}) dia
 }
 
 func dataToUserRequest(data *schema.ResourceData) (req fusionauth.UserRequest, diags diag.Diagnostics) {
-	twoFactorMethods, diags := dataToTwoFactorMethods(data)
-	if diags != nil {
-		return
+	twoFactorMethods, subDiags := dataToTwoFactorMethods(data)
+	if subDiags != nil {
+		diags = append(diags, subDiags...)
 	}
 
-	userData, diags := jsonStringToMapStringInterface("data", data.Get("data").(string))
-	if diags != nil {
-		return
+	resourceData, subDiags := jsonStringToMapStringInterface("data", data)
+	if subDiags != nil {
+		diags = append(diags, subDiags...)
 	}
 
 	req = fusionauth.UserRequest{
 		User: fusionauth.User{
 			TenantId:           data.Get("tenant_id").(string),
 			BirthDate:          data.Get("birth_date").(string),
-			Data:               userData,
+			Data:               resourceData,
 			Email:              data.Get("email").(string),
 			Expiry:             int64(data.Get("expiry").(int)),
 			FirstName:          data.Get("first_name").(string),
@@ -253,54 +252,6 @@ func userResponseToData(data *schema.ResourceData, resp *fusionauth.UserResponse
 	}
 
 	return nil
-}
-
-// jsonStringToMapStringInterface takes in a json encoded string and transforms
-// the data to a map[string]interface{} to comply to the expected type for the
-// fusionauth client.
-func jsonStringToMapStringInterface(fieldName string, in string) (out map[string]interface{}, diags diag.Diagnostics) {
-	out = map[string]interface{}{}
-	if strings.TrimSpace(in) == "" {
-		return out, nil
-	}
-
-	if err := json.Unmarshal([]byte(in), &out); err != nil {
-		diags = append(diags, diag.Diagnostic{
-			Severity: diag.Error,
-			Summary:  fmt.Sprintf("Unable to transform %s to expected type", fieldName),
-			Detail: fmt.Sprintf(
-				"Error unmarshalling %s from an expected JSON encoded string to map[string]interface{}.\n"+
-					"Please make sure you have wrapped your HCL with jsonencode."+
-					"For example, 'jsonencode({ hello = \"world\" })'.\n\n"+
-					"error: %s\n",
-				fieldName, err,
-			),
-		})
-	}
-
-	return
-}
-
-func mapStringInterfaceToJSONString(fieldName string, in map[string]interface{}) (out string, diags diag.Diagnostics) {
-	if len(in) == 0 {
-		return "", nil
-	}
-
-	outBytes, err := json.MarshalIndent(in, "", "  ")
-	if err != nil {
-		diags = append(diags, diag.Diagnostic{
-			Severity: diag.Error,
-			Summary:  fmt.Sprintf("Unable to transform %s to expected type", fieldName),
-			Detail: fmt.Sprintf(
-				"Error marshalling %s from a map[string]interface{} to a JSON string.\n"+
-					"error: %s\n",
-				fieldName, err,
-			),
-		})
-		return
-	}
-
-	return string(outBytes), nil
 }
 
 func dataToTwoFactorMethods(data *schema.ResourceData) (twoFactorMethods []fusionauth.TwoFactorMethod, diags diag.Diagnostics) {
